fix some return type errors
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@31338 72102866-910b-0410-8b05-ffd578937521

+++ Source/GSAttributedString.m
@@ -512,7 +512,7 @@ _attributesAtIndexEffectiveRange(
 
 
 // The superclass implementation is correct but too slow
-- (unsigned int) length
+- (NSUInteger) length
 {
   return [_textChars length];
 }
@@ -901,7 +901,7 @@ SANITY();
 }
 
 // The superclass implementation is correct but too slow
-- (unsigned int) length
+- (NSUInteger) length
 {
   return [_textChars length];
 }
